requirements.txt
Flask==3.0.3
mlflow==2.15.0
python-dotenv==1.0.1